Make build windows compatible

+++ src/test/scala/com/microsoft/ml/spark/downloader/DownloaderSuite.scala
@@ -3,9 +3,9 @@
 
 package com.microsoft.ml.spark.downloader
 
+import java.io.File
 import java.nio.file.Files
 
-import com.microsoft.ml.spark.core.env.FileUtilities.File
 import com.microsoft.ml.spark.core.test.base.TestBase
 import org.apache.commons.io.FileUtils
 
@@ -69,7 +69,9 @@ class DownloaderSuite extends TestBase {
   }
 
   override def afterAll(): Unit = {
-    FileUtils.forceDelete(saveDir)
+    if (saveDir.exists()) {
+      FileUtils.forceDelete(saveDir)
+    }
     super.afterAll()
   }
